# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: AISI GROUP
  description: API for managing property
  version: 0.2.0
servers:
  - url: http://localhost:3001/api
    description: Local server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: 'Error message'
...
paths:
...
  /add_building:
    post:
      tags:
        - Buildings
      summary: 'Add a new building'
      description: 'Create a new building with only a name'
      operationId: 'addBuilding'
      security:
        - cookieAuth: []
      requestBody:
        description: 'Building data including only the name'
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Goderdzi Resort'
                  description: 'The name of the building'
              required:
                - name
      responses:
        '201':
          description: 'Building created successfully'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Building created successfully'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'b1d8fd3a-7ab1-4f1e-8353-e4f98e5c094c'
                      name:
                        type: string
                        example: 'Goderdzi Resort'
                      created_at:
                        type: string
                        format: date-time
        '400':
          description: 'Invalid input data'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Invalid input data'
                  errors:
                    type: object
                    description: 'Validation error details'
        '401':
          description: 'Authentication required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'Invalid or expired token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Internal server error while creating Building'
  /delete_building/{id}:
    delete:
      tags:
        - Buildings
      summary: 'Delete a building'
      description: 'Delete a building by its ID'
      operationId: 'deleteBuilding'
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 'The ID of the building to delete'
          schema:
            type: string
      responses:
        '200':
          description: 'Building deleted successfully'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Building deleted successfully'
        '404':
          description: 'Building not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'Authentication required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'Invalid or expired token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /building/{id}:
    get:
      tags:
        - Buildings
      summary: 'Get building by ID'
      description: 'Retrieve a building by its ID'
      operationId: 'getBuildingById'
      parameters:
        - name: id
          in: path
          required: true
          description: 'The ID of the building to retrieve'
          schema:
            type: string
      responses:
        '200':
          description: 'Building retrieved successfully'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Building retrieved successfully'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'b1d8fd3a-7ab1-4f1e-8353-e4f98e5c094c'
                      name:
                        type: string
                        example: 'Goderdzi Resort'
                      created_at:
                        type: string
                        format: date-time
        '404':
          description: 'Building not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /update_apartment_status:
    put:
      summary: Update the status of an apartment
      description: Updates the status of an apartment based on the floor plan ID and flat number.
      operationId: updateApartmentStatus
      tags:
        - Apartments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - available
                    - reserved
                    - sold
                  description: The new status of the apartment.
                floor_plan_id:
                  type: integer
                  description: The ID of the floor plan that the apartment belongs to.
                flat_number:
                  type: integer
                  description: The flat number of the apartment.
              required:
                - status
                - floor_plan_id
                - flat_number
      responses:
        '200':
          description: Apartment status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Apartment status updated successfully'
                  apartment:
                    type: object
                    properties:
                      flat_id:
                        type: integer
                        example: 1
                      flat_number:
                        type: integer
                        example: 101
                      status:
                        type: string
                        example: 'reserved'
                      floor_plan_id:
                        type: integer
                        example: 8
                      image:
                        type: string
                        nullable: true
                        example: null
                      square_meters:
                        type: string
                        example: '50.00'
        '400':
          description: Invalid input. The request body is not valid or missing required fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Invalid input'
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        message:
                          type: string
                        path:
                          type: array
                          items:
                            type: string
                        type:
                          type: string
        '404':
          description: No apartment found with the provided floor_plan_id and flat_number.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Apartment not found with the given floor_plan_id and flat_number'
        '500':
          description: Internal server error. Something went wrong on the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Internal server error'
...